feat: Criação da Pasta Code
