--- README.md
@@ -1,65 +1,0 @@
-### Objetivo do Projeto:
-
-O projeto tem como objetivo desenvolver um modelo de linguagem natural capaz de distinguir se um texto foi gerado por uma Linguagem de Modelo de Máquina (LLM) ou por uma pessoa. O sistema buscará atingir uma classificação precisa, adaptando-se a diferentes arquiteturas de modelos de linguagem e sendo capaz de lidar com novos padrões de linguagem ao longo do tempo.
-
-Os requisitos funcionais incluem a implementação de algoritmos de aprendizado de máquina para análise de características distintivas, adequação a diferentes modelos de linguagem, adaptabilidade a novos dados e desempenho em diversos idiomas. Além disso, o sistema deve processar dados não estruturados e manter uma alta precisão na classificação.
-
-Adicionalmente, considera-se a possibilidade de suporte a múltiplos domínios e a identificação de técnicas específicas de geração de LLM, bem como a implementação de um sistema de feedback para melhorias contínuas.
-
-O projeto visa criar um modelo robusto, escalável e eficiente, cumprindo regulamentações éticas e de privacidade, com a capacidade de treinamento contínuo para aprimoramento ao longo do tempo.
-
-### Requisitos Funcionais:
-
-1. **Classificação Precisa:**
-   - O sistema deve empregar algoritmos de aprendizado de máquina capazes de analisar características distintivas entre textos gerados por linguagem natural e textos gerados por modelos de linguagem de máquina. A precisão na classificação deve ser medida por métricas como taxa de verdadeiros positivos, verdadeiros negativos, falsos positivos e falsos negativos.
-
-2. **Adequação a Diferentes Modelos:**
-   - O modelo deve ser projetado de maneira flexível para suportar diferentes arquiteturas de modelos de linguagem, garantindo que a análise seja eficaz independentemente do tipo específico de LLM utilizado.
-
-3. **Adaptabilidade a Novos Dados:**
-   - O sistema deve ser capaz de aprender continuamente e se adaptar a novos padrões de linguagem que possam surgir. Isso pode envolver técnicas de treinamento online e atualização periódica do modelo.
-
-4. **Desempenho em Diferentes Idiomas:**
-   - O modelo deve ser treinado e testado em uma variedade de idiomas para garantir que a classificação seja robusta em contextos multilíngues.
-
-5. **Manuseio de Dados Não Estruturados:**
-   - O sistema deve ser capaz de processar textos de diferentes comprimentos e formatos, incluindo dados não estruturados, como conversas informais e textos coloquiais.
-### Requisitos Funcionais Adicionais:
-
-6. **Identificação de Técnicas de Geração:**
-   - Além de distinguir entre LLM e texto humano, o sistema pode ser expandido para identificar características específicas associadas a técnicas de geração de LLM, como padrões de repetição, coesão textual e estruturas gramaticais específicas.
-
-7. **Feedback para Melhorias:**
-   - Implementar um sistema de feedback que permite aos usuários fornecerem correções ou indicarem instâncias em que a classificação do modelo pode ser aprimorada, contribuindo assim para melhorias contínuas.
-
-8. **Suporte a Múltiplos Domínios:**
-   - O modelo pode ser estendido para lidar com textos em diferentes domínios (por exemplo, notícias, literatura técnica, redes sociais), garantindo uma aplicabilidade mais ampla.
-
-### Requisitos Não Funcionais:
-
-1. **Desempenho:**
-   - O modelo deve fornecer uma resposta rápida, com tempos de inferência minimizados para suportar aplicações em tempo real.
-
-2. **Confiança e Interpretabilidade:**
-   - O sistema deve fornecer explicações claras sobre as razões por trás de suas classificações, facilitando a compreensão e aumentando a confiança dos usuários na validade das decisões do modelo.
-
-3. **Segurança:**
-   - O modelo deve ser resistente a ataques adversários e manipulações, garantindo a integridade e a confiabilidade das decisões de classificação.
-
-4. **Escalabilidade:**
-   - O sistema deve ser capaz de dimensionar eficientemente para lidar com grandes volumes de dados e cargas de usuários crescentes sem comprometer o desempenho.
-
-5. **Manutenibilidade:**
-   - O código-fonte e os artefatos do modelo devem ser bem documentados e organizados, facilitando futuras atualizações, correções e manutenção.
-
-6. **Privacidade:**
-   - Garantir que o modelo respeite a privacidade dos usuários, evitando a retenção desnecessária de informações sensíveis e aplicando práticas de segurança adequadas.
-
-7. **Facilidade de Integração:**
-   - O modelo deve ser facilmente integrável a sistemas existentes por meio de APIs ou outras interfaces, permitindo uma adoção suave em ambientes já estabelecidos.
-
-8. **Treinabilidade Contínua:**
-   - Permitir a atualização do modelo com novos dados para melhorar a precisão ao longo do tempo, incorporando mecanismos de treinamento contínuo.
-
-9. **Interpretação Humana:**
-   - Certificar-se de que as decisões do modelo são apresentadas de maneira que seja compreensível para usuários humanos, promovendo transparência e interpretabilidade.
